requirements.txt
alembic==1.13.1
asyncpg==0.29.0
bcrypt==4.1.3
fastapi==0.111.0
Jinja2==3.1.4
progressbar2==4.4.2
psycopg2-binary==2.9.9
pydantic-settings==2.2.1
pydantic[email]==2.3.*
pydantic_core==2.6.3
pydantic==2.3.0
redis==5.0.4
ruff==0.4.4
SQLAlchemy==2.0.30
uvicorn==0.29.0
httpx==0.24.*



starlette~=0.37.2
python-dotenv~=1.0.1
python-dateutil~=2.9.0.post0